feat: add support for retrieving product version from group config
In case the build config used to build the project doesn't have the
`productVersion` set, we will try to get this information from a group
configuration this build config is assigned to (if any).

In case the build config is assigned to multiple group configs -- we
will take into account all of them. This will result in multiple
"products" in SBOMer terms which will rpoduce multiple SBOMs for all of
the productVersions found.