feat: support for providing configuration for generation
This change makes it possible to provide the Config with the product and
generator configuration as part of the REST API call to
/api/v1alpha1/sboms/generate/build/{buildId}. It enables providing a
custom configuration for that particular generation.

In case the configuration is provided it will be used, which means that
the initialziation phase which includes crafrting configuration won't be
run at all.

This feature is very useful for cases where we want to try out different
configurations to "get it right", for example when onboarding someone.
Another use case is to adjust configuration so that SBOM can be
generated for an old build.

Configuration can be sent in the body as YAML (application/yaml
content type) or JSON (application/json content type). Provided
configuration is validated and adjusted if necessary (putting default
values, if not provided).