feat(service): NCL-8201: exposure of /sboms/purl/[PURL] endpoint
This new endpoint makes it possible to retrieve the SBOM by giving just
the purl. In case of multiple SBOMs will be found for a given purl,
latest SBOM will be returned.

In case no SBOM can be found, 404 will be returned.

Additionally `/sboms/purl/[PURL]/bom` is exposed as well to retrieve the
BOM content.