feat(SBOMER-16): do not bundle tools
This is an attempt to make the generator image smaller and more
agile. Instead of bundling all the necessary tools and SDKs at build
time, we do it selectively at runtime based on the environment
information as well as config.

The only thing we install at build time is: JDK 17 which is required to
run SBOMer CLI. This will prpobably change in the future once we move to
natives.

We install nvm as well sdkman to be able to install required tools at
build time.

Fixes: https://issues.redhat.com/browse/SBOMER-16