feat(SBOMER-14): open generation for all builds
This commit makes it possible to generate manifests for any PNC builds,
no matter if these are related to a produc, or not.

This required adjustments in the validation, because now the Red Hat
product processor is optional.

Additionally this moves the logic for environment configuration
preparation into the initialziation phase. Removes the env-detection
phase entirely.

https://issues.redhat.com/browse/SBOMER-15
https://issues.redhat.com/browse/SBOMER-14