fix(SBOMER-14): always generate environment config if missing
This fixes a problem for cases where the configuration if provided by
the internal SBOMer mapping or via configuration stored in the project's
repository. In such cases we won't have most probably the environment
field populated.

The logic was changed so that we will generate the environment always in
case it is not provided already.

Adiditional cleanup of the code was performed as well.