feat(SBOMER-51): use the Jakarta-supported version of PNC REST library
Earlier we had to add many of the PNC REST library, because this was the
easiest way to just have it to be able to upgrade to Quarkus 3 which
switched from javax to jakarta. Now, since we have a dedicated artifact
from the PNC team to support Jakarta we can get rid of these.