ci: image building and deployment on GitLab
This basically replicates the GitHub Actions in GitLab Pipelines.

Remove the buildnumber-maven-plugin. It's unnecessary, locally we set up
the build number to `dev` and when building on CI we set it to Git short
hash.

Introduction of custom `settings.xml` file that points to a NGINX cache
for Maven repositories.