feat: add support for feature-flags
Initial implementation adds support for a single `dry-run` toggle.
When enabled, no changes will be allowed to be made via the REST API
returning 405 HTTP code.

Support for disabling UMB processing will be added later, because it is
non-trivial to disable/enable it on demand.

In dev mode, feature flags are disabled entirely and default values from
`FeatureFlags` class will be used.