fix(SBOMER-69): use specific version of Java to run Domino
Recently Domino upgraded Java to 17, which means that we cannot use it
with older JDKs, but at the same time we try to be as close as we can
be to the build environment. But in thiscase this doesn't hurt to use
a newer JDK. This makes use of JDK 17 for Domino generator.

Fixes: https://issues.redhat.com/browse/SBOMER-69