feat(SBOMER-8): implement support for storing files in S3
It adds a s3 client via facade so that we can enable or disable the
client based on the feature flag. When it is enabled, an event is sent
and the client is initialized. When it is disabled, the client is
closed.

Names of generation Tekton tasks were adjusted so that we can have a
predictable name. It uses the generation request ID from now on.

https://issues.redhat.com/browse/SBOMER-8