feat(SBOMER-83): support for syft generator
This does add support for the Syft generator. Currently it's usage is
limited to manifesting products located within container images.

Notable changes:

Default processor was moved out of command and it is now a separate
entity. This makes it possible to use it across different commands and
not only.

The default processor is now processing every component found in the
manifest, not only the ones with `-redhat` suffix. This will increase
the time of generation, but we should get even more accurate results.
Other change is that If we cannot find given purl within PNC, we will be
using hashes (if available) to lookup artifacts.

The container image manifest generator is capable of generating
manifests for Application Services products that are shipped in
container images. Currently it support only products that are located
under the `/opt` directory. In the future this will be configurable.

The `redhat` processor is not applied for these manifests for now. It
will be enabled at a later stage.

This change introduces "Adjuster" concept. Adjuster is a processor that
runs after generation and processes the outout of the generator to a
format that is understandable by SBOMer. Adjuster can do some light
tweaks or entirely rewrite the manifest. Main use case if to enable use
of external generators. In this specific case we have the Syft
generator. It's output is processed, properties are renamed and moved.

Introduction of adjuster and standa-alone generators required to change
the CLI of SBOMer. Now instead of running the generation and then
processing we run adjustment nad processing on an already-existing
manifest which is provided by an external generator.

https://issues.redhat.com/browse/SBOMER-83