fix(SBOMER-86): handle deliverable analyzer failures
This makes it possible to handle messages from PNC for operations that
have failed. Previously we kind of assumed these will be successfull,
but we were proved wrong very fast.

When we receive a failure message, we update the Generation request
entity as well and set the proper status with a reason pointing what
went wrong.

https://issues.redhat.com/browse/SBOMER-86