refactor: generalize Generation configuration
This change makes implementing different types of configurations much easier. Configs are different for different generation types. In the early days we had just one related to PNC build, but now we have 4, which will probably rise in the future as well.

In order to make implementation of newa features easier configuration handling had to be reworked. A Config class with helper methods was introduced and now all configs are derived from it. This makes serialization and deserialization much easier. Besides this we can now operatore on the object level and not JsonNode within entities.