feat: umb notification improvements
Updated documentation to inform how to request UMB notification resend.

Added a test for the notification resend. While working on it it become evident, that we expose it only for v1alpha2 version. This was fixed and it is now available in all endpoints.

Additionally notification erros are throwing exceptions. Earlier we could easily miss them.