test: update e2e tests after API refactoring
All e2e tests had to be updated, because latest API version returns
array of generations in response.

Reading of config files was adjusted because we do not default to the
PncBuildConfig.

One image test was removed, because it's the same as the other one (we
do not allow to provide product - errata - config anymore).