chore: improve Quarkus testing feedback loop
Some background: Quarkus is restarted whenever a test resource is
restricted to the class or different from the ones from other tests and
also when the test profile is different.

There is a clever orderer in Quarkus that tries to order things cleverly
to minimize the number of restarts but... ITs and Failsafe are not
honoring the JUnit test orderer so you lose this benefit and you end up
with Quarkus being restarted a lot more often than strictly necessary.

So the first fix here is to consider the Quarkus tests as Surefire
tests, not failsafe tests.
If more practical for you, they can be isolated in a specific Surefire
execution but we should keep them executed with Surefire.

Second issue is that @WithKubernetesTestServer is restricted to a given
test clean, meaning every time you have this annotation, the server will
have to be restarted.
I will start discussions to see if we should change that but in the
meantime I added an annotation that has a less aggressive behavior.

Unfortunately, you are also affected by a Quarkus bug, for which I will
push a fix soon and that we will hopefully backport to the next 3.17.