chore: update to latest Quarkus micro
While I agree you should stay compatible with the minor requested by the
Operator SDK extension, I don't think it's a good idea to stay on the
micro as you are missing all the fixes we are pushing since at the
moment the version used here is 3.17.0.

So my advice is: update to the latest micro of the minor branch
requested for the Operator SDK.

This is important for #1076 as otherwise, you will have to wait a lot
longer to get the fix. Also it's good practice in general.