feat(SBOMER-81): Add license information to components
If given an `AnalyzedArtifact`, add any SPDX license identifers to the
licenses of the component. Only licenses for which an SPDX license
identifier assertion could be made are listed here.

Since URLs are not unique per identifier, do not add any URL to the
licenses of the component. Add any valid absolute URLs to the external
references of the component.

Add the full list of licenses, including any licenses which have an
SPDX license identifier of `NOASSERTION` or `NONE` to the license
evidence of the component. The evidence licenses include URL and
source information.