feat(service): adjust handling of new generations
New generation do not automatically create ConfigMap representations in the cluster
which received the event. Instead, an entity in the database is created
and periodically synced with the cluster.

For this purpose a `GenerationRequestScheduler` bean was created. It's
responsibility is to look periodically in the database for generations
with the NEW state and created representations in the cluster (as CM).
When a generation is synced with cluster its status is changed to
`SCHEDULED`.

Syncing is done periodically. Interval by default is set to 15 seconds.
Scheduler starts with a delay of one minute after the service is up.

In order to prevent creation of duplicates resources in the cluster
running multiple replicas `LeaderManager` was introduced. It controls
and designates a single leader within a cluster. It checks the status of
the lease every 10 seconds by default and if there is no leader or the
conditions to become new leader are met -- it changes current instance
to become a leader.

The `GenerationRequestScheduler` works only on the instance which is a
leader.

Side effect of all this change is that no matter which cluster created
the generation request -- all active clusters will be equally
responsible for handling generations. This means that it should
potentially spread the load across AZ as well.

Notable changes:

- Adds a new `SCHEDULED` status to generation. It's main purpose is to
  differentiate between the ones that are strictly new (just created)
  from the onese that are synced with some cluster.

Refs: SBOMER-387