chore: introduction of resolvers (/events/resolve API)
This makes use of DTO's more heavily. But most important part is the
introduction of POST `/api/v1beta2/events/resolve` endpoint which makes
use of resolvers.

Resolver are used when we have some identifier and we know that this
indetifier is related to some service. For example Errata Tool or GitHub
or Jenkins.

We can request "handling" of a particular event by providing the
resolver name (type) and the identifier. This will create a new event
within the system and the resolver will be responsible for getting more
information about the work that needs to be done (deliverables to
manifest).

This commit contains an example of RedHatAdvisoryEventListener which
handles Errata Tool advisories. This is dummy code for now.
The important part is that it is handled in an async way and then it
creates generations and updates the event entity.

The goal is that it should happen by calling a secured REST API, but for
now (when we colocate the handler and service) this is ok.